compliancegpt/backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
httpx>=0.28.1
beautifulsoup4>=4.12.3
google-genai>=1.14.0
supabase>=2.4.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1
apscheduler>=3.10.4
resend>=0.7.2
pypdf2>=3.0.1
python-dotenv>=1.0.1
pydantic>=2.7.1
python-multipart>=0.0.9
